requirements.txt
# Python version
# Python==3.11.0

streamlit==1.51.0
opencv-python==4.12.0.88
mediapipe==0.10.21
numpy==1.26.4
google-genai==1.49.0